requirements.txt
openai~=2.13.0
langchain~=1.2.0
ssim~=0.3.0
numpy~=1.26.4
torch~=2.9.1
scikit-image~=0.25.2
pillow~=12.0.0
scipy~=1.15.3
tqdm~=4.67.1
opencv-python~=4.12.0.88
langgraph~=1.0.5